Added endpoint to send the user in the session to front end

+++ src/main/java/com/revature/controllers/ProductController.java
@@ -83,15 +83,20 @@ public class ProductController {
 
     @Authorized
     @DeleteMapping("/{id}")
-    public ResponseEntity<Product> deleteProduct(@PathVariable("id") int id) {
+    public ResponseEntity<Product> deleteProduct(@PathVariable("id") int id, HttpSession session) {
         Optional<Product> optional = productService.findById(id);
+        User u= (User) session.getAttribute("user");
+        if (u.getRole().toString()=="Admin") {
+            if (!optional.isPresent()) {
+                return ResponseEntity.notFound().build();
+            }
+            productService.delete(id);
 
-        if(!optional.isPresent()) {
+            return ResponseEntity.ok(optional.get());
+        }
+        else{
             return ResponseEntity.notFound().build();
         }
-        productService.delete(id);
-
-        return ResponseEntity.ok(optional.get());
     }
 
     @Authorized
